Issue 12024: NB! Current build is broken - work in progress. Replaced user loading using SQLs with JPA
git-svn-id: https://svn.imcode.com/imcode/imcms/trunk@30631 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ src/main/java/com/imcode/imcms/addon/imagearchive/service/ImageService.java
@@ -621,7 +621,7 @@ public class ImageService {
                 }
             }
 
-            List<Integer> keywordIds = new ArrayList<>(newKeywordNames.size() + existingKeywords.size());
+            List<Long> keywordIds = new ArrayList<>(newKeywordNames.size() + existingKeywords.size());
             for (String k : newKeywordNames) {
                 Keywords keyword = new Keywords();
                 keyword.setKeywordNm(k);
@@ -653,7 +653,7 @@ public class ImageService {
                         .executeUpdate();
             }
 
-            for (Integer id : keywordIds) {
+            for (Long id : keywordIds) {
                 ImageKeywords ik = new ImageKeywords();
                 ik.setKeywordId(id);
                 ik.setImageId(imageId);
